refactor: remove web response in jwt
