+++ app/main.py
@@ -11,3 +11,4 @@ app = FastAPI()
 
 app.include_router(auth.router)
 app.include_router(jwt.router)
+
