+++ app/models/dto/token_base_model.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+    
+class TokenData(BaseModel):
+    name: str | None = None
