+++ app/models/dto/user_base_model.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class User(BaseModel):
+    name: str
+
+class UserInDB(User):
+    password: str
